+++ packages/services/src/index.ts
@@ -12,3 +12,4 @@ export * from './lens';
 export * from './mind';
 export * from './ports';
 export * from './sdk';
+export * from './tools';